Spider Quotes Scraper 2 v 1.3.0.2

+++ quotes_scraper/quotes_scraper/quotes_scraper/spiders/quotes.py
@@ -17,7 +17,9 @@ class quotesSpider(scrapy.Spider):
     start_urls = [
         'http://quotes.toscrape.com/'
     ]
-    os.remove("quotes.json")
+    if os.path.exists('quotes.json'):
+        os.remove("quotes.json")
+    
     custom_settings = {
         "FEEDS":{"quotes.json":{"format":"json"}}
     }
